Add incentiveInfo to garData
