+++ src/lib/convert.ts
@@ -1,12 +1,15 @@
-import {
-  zGatewayAddressRegistryCache,
-  zGatewayAddressRegistryItem,
-} from "@/types";
+import { zGatewayAddressRegistryItem } from "@/types";
 import { z } from "zod";
 import { linkFull, linkDisplay } from "./utils";
+import {
+  incentiveContractSchema,
+  incentiveInfoSchema,
+} from "./incentive/schema";
+import { zGatewayAddressRegistryCache } from "./gar/schema";
 
 const extractGarItems = (
-  garCache: z.infer<typeof zGatewayAddressRegistryCache>
+  garCache: z.infer<typeof zGatewayAddressRegistryCache>,
+  incentiveContractData?: z.infer<typeof incentiveContractSchema>
 ): Array<z.infer<typeof zGatewayAddressRegistryItem>> => {
   const gatewayEntries = Object.entries(garCache.gateways);
   return gatewayEntries.map(([txId, item]) => {
@@ -17,6 +20,18 @@ const extractGarItems = (
       fqdnIndex === 0
         ? item.settings.fqdn
         : `${item.settings.fqdn}${`~${fqdnIndex}`}`;
+
+    let incentiveInfo: z.infer<typeof incentiveInfoSchema> | undefined;
+    // TODO: find a better way to match these up
+    const incentiveItem = incentiveContractData?.[txId];
+    if (incentiveItem) {
+      incentiveInfo = {
+        stats: incentiveItem.stats,
+        weights: incentiveItem.weights,
+      };
+    }
+    if (incentiveInfo != undefined) console.log("incentiveItem", incentiveItem);
+
     return {
       id: txId,
       fqdnKey: fqdnKey,
@@ -34,6 +49,7 @@ const extractGarItems = (
         item.settings.port
       ),
       ...item,
+      incentiveInfo,
     };
   });
 };
