+++ src/lib/incentive/schema.ts
@@ -1,26 +1,28 @@
-import { zArweaveTxId, zGatewayAddressRegistryItemData } from "@/types";
 import { z } from "zod";
+import { zArweaveTxId, zGatewayAddressRegistryItemData } from "../gar/schema";
 
-export const incentiveContractItemSchema = z.union([
-  zGatewayAddressRegistryItemData,
-  z.object({
-    stats: z.object({
-      failedConsecutiveEpochs: z.number().int().nonnegative(),
-      passedEpochCount: z.number().int().nonnegative(),
-      submittedEpochCount: z.number().int().nonnegative(),
-      totalEpochParticipationCount: z.number().int().nonnegative(),
-      totalEpochsPrescribedCount: z.number().int().nonnegative(),
-    }),
-    weights: z.object({
-      compositeWeight: z.number().nonnegative(),
-      gatewayRewardRatioWeight: z.number().nonnegative(),
-      normalizedCompositeWeight: z.number().nonnegative(),
-      observerRewardRatioWeight: z.number().nonnegative(),
-      stakeWeight: z.number().nonnegative(),
-      tenureWeight: z.number().nonnegative(),
-    }),
+export const incentiveInfoSchema = z.object({
+  stats: z.object({
+    failedConsecutiveEpochs: z.number().int().nonnegative(),
+    passedEpochCount: z.number().int().nonnegative(),
+    submittedEpochCount: z.number().int().nonnegative(),
+    totalEpochParticipationCount: z.number().int().nonnegative(),
+    totalEpochsPrescribedCount: z.number().int().nonnegative(),
   }),
-]);
+  weights: z.object({
+    compositeWeight: z.number().nonnegative(),
+    gatewayRewardRatioWeight: z.number().nonnegative(),
+    normalizedCompositeWeight: z.number().nonnegative(),
+    observerRewardRatioWeight: z.number().nonnegative(),
+    stakeWeight: z.number().nonnegative(),
+    tenureWeight: z.number().nonnegative(),
+  }),
+});
+
+export const incentiveContractItemSchema = z.intersection(
+  zGatewayAddressRegistryItemData,
+  incentiveInfoSchema
+);
 
 export const incentiveContractSchema = z.record(
   zArweaveTxId,
