+++ src/lib/query.ts
@@ -5,15 +5,20 @@ import {
   downloadReportInfoForTransaction,
   querySingleTransaction,
 } from "./observer/downloadObservation";
+import { fetchIncentiveContractData } from "./incentive/fetchIncentiveContractData";
+import { zGatewayAddressRegistryCache } from "./gar/schema";
 
 export const queryClient = new QueryClient();
 
 export const garQuery = queryOptions({
   queryKey: ["gar"],
   queryFn: async () => {
-    const fetchResult = await fetch(defaultGARCacheURL);
-    const fetchJson = await fetchResult.json();
-    const garItems = extractGarItems(fetchJson);
+    const garCache = await fetch(defaultGARCacheURL)
+      .then((res) => res.json())
+      .then((json) => zGatewayAddressRegistryCache.parse(json));
+    const incentiveContractData = await fetchIncentiveContractData();
+
+    const garItems = extractGarItems(garCache, incentiveContractData);
     garItems.sort((a, b) => a.fqdnKey.localeCompare(b.fqdnKey));
     return garItems;
   },
